Title: 이차원 배열의 오름차순 정리, Time: 96ms, Memory: 9MB, Status: Passed - CodeTree

+++ 250207/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.java
@@ -3,13 +3,13 @@ import java.io.*;
 
 public class Main {
     
-    static int N, K;
+    static long N, K;
 
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
-        N = Integer.parseInt(br.readLine());
-        K = Integer.parseInt(br.readLine());
+        N = Long.parseLong(br.readLine());
+        K = Long.parseLong(br.readLine());
 
         long left = 1;
         long right = N * N;
